Stipple.Elements.root — Function

function root(app::M)::String where {M<:ReactiveModel}

Generates a valid JavaScript object name to be used as the name of the Vue app – and its respective HTML container.

Stipple.Elements.elem — Function

function elem(app::M)::String where {M<:ReactiveModel}

Generates a JS id # reference to the DOM element containing the Vue app template.

Stipple.Elements.vm — Function

function root(app::M)::String where {M<:ReactiveModel}

Generates a valid JavaScript object name to be used as the name of the Vue app – and its respective HTML container.

Stipple.Elements.vue_integration — Function

function vue_integration(model::M; vue_app_name::String, endpoint::String, debounce::Int)::String where {M<:ReactiveModel}

Generates the JS/Vue.js code which handles the 2-way data sync between Julia and JavaScript/Vue.js. It is called internally by Stipple.init which allows for the configuration of all the parameters.

Stipple.Elements.@if — Macro

@if(expr)

Generates v-if Vue.js code using expr as the condition. https://vuejs.org/v2/api/#v-if

Example

julia> span("Bad stuff's about to happen", class="warning", @if(:warning))
"<span class="warning" v-if='warning'>Bad stuff's about to happen</span>"

Tentatively we now also support Julia expressions with comparison operators ```julia julia> cell(@if(:i ∉ 3:2:7)) "<div class="st-col col" v-if="!(i in [3,5,7])"></div>"

julia> row("hello", @showif(:n^2 ∉ 3:2:11)) "<div v-show="!(n ** 2 ∉ [3,5,7,9,11])" class="row">hello</div>" ````

Stipple.Elements.@elseif — Macro

@elseif(expr)

Generates v-else-if Vue.js code using expr as the condition. https://vuejs.org/v2/api/#v-else-if

Example

julia> span("An error has occurred", class="error", @elseif(:error))
"<span class="error" v-else-if='error'>An error has occurred</span>"

Stipple.Elements.@else — Macro

@else(expr)

Generates v-else Vue.js code using expr as the condition. https://vuejs.org/v2/api/#v-else

Example

julia> span(@else, "Might want to keep an eye on this", class="notice")
"<span v-else class="notice">Might want to keep an eye on this</span>"

Stipple.Elements.@for — Macro

Generates v-for directive to render a list of items based on an array. https://vuejs.org/v2/guide/list.html#Mapping-an-Array-to-Elements-with-v-for

@for supports both js expressions as String or a Julia expression with Vectors or Dicts

Example

Javascript

julia> p(" {{todo}} ", class="warning", @for("todo in todos"))
"""
<p v-for='todo in todos'>
    {{todo}}
</p>
"""

Julia expression

julia> dict = Dict(:a => "b", :c => 4);
julia> ul(li("k: {{ k }}, v: {{ v }}, i: {{ i }}", @for((v, k, i) in dict)))
"""
<ul>
    <li v-for="(v, k, i) in {'a':'b','c':4}">
        k: {{ k }}, v: {{ v }}, i: {{ i }}
    </li>
</ul>
"""

Note the inverted order of value, key and index compared to Stipple destructuring. It is also possible to loop over (v, k) or v; index will always be zero-based

Stipple.Elements.@text — Macro

@text(expr)

Creates a v-text or a text-content.prop Vue biding to the element's textContent property. https://vuejs.org/v2/api/#v-text

Example

julia> span("", @text("abc | def"))
"<span :text-content.prop='abc | def'></span>"

julia> span("", @text("abc"))
"<span v-text='abc'></span>"

Stipple.Elements.@bind — Macro

@bind(expr, [type])

Binds a model parameter to a Vue component, generating a v-model property, optionally defining the parameter type. https://vuejs.org/v2/api/#v-model

Example

julia> input("", placeholder="Type your name", @bind(:name))
"<input placeholder="Type your name"  v-model='name' />"

julia> input("", placeholder="Type your name", @bind(:name, :identity))
"<input placeholder="Type your name"  v-model.identity='name' />"

Stipple.Elements.@data — Macro

@data(expr)

Creates a Vue.js data binding for the elements that expect it.

Example

julia> plot(@data(:piechart), options! = "plot_options")
"<template><apexchart :options="plot_options" :series="piechart"></apexchart></template>"

Stipple.Elements.@on — Macro

`@on(action, expr)`

Defines a js routine that is called by the given action of the Vue component, e.g. :click, :input

Example

julia> input("", @bind(:input), @on("keyup.enter", "process = true"))
"<input  v-model='input' v-on:keyup.enter='process = true' />"

If expr is a symbol, the event will be sent to the backend. In order to handle the event a respective Stipple.notify method needs to be defined either manually or via the @event macro.

Example

julia> Stipple.notify(model, ::Val{:my_click}) = println("clicked")

or if event information is needed

Stipple.notify(model, ::Val{:my_click}, event_info) = println(event_info)

or via the @event macro:

@event :my_click begin
  @info "clicked, event_info is " event
  notify(__model__, "Info from the backend: Clicked")
end

Note that in the handler model refers to the receiving model and event is a Dict of event information. The handler is linked in the ui-element

btn("Event test", @on("click", :my_click))

Sometimes preprocessing of the events is necessary, e.g. to add or skip information

@on(:uploaded, :uploaded, "for (f in event.files) { event.files[f].fname = event.files[f].name }")

This is necessary because in some cases, e.g. in case of the click event not all fields are automatically converted by JSON.stringify. Other events, e.g. the row-click event of the q-table component pass more arguments than just the event itself. These arguments are accessible as args.

table(:table, @on(:row__click, :rowclick, "event.row = args[0]"))

You can also use predefined handlers as preprocessors. This is useful if you want to use the same handler for multiple events. Handler names are passed as Symbols.

@methods [
    :addTableIndex => js"""
        function (event, row, id) {
            const keys = Object.keys(row)
            event.row = id + 1;
            event.column = event.target.closest('td').cellIndex + 1;
            event.value = row[keys[event.column]];
            event.row_data = row;
            event.column_keys = keys;
            return event;
        }
    """
]

cell(@on(:click, :rowclick, :addTableIndex))

Finally, you can also pass arrays of handlers as preprocessors, make sure in this case that each handler returns the event object, as the handlers are chained. Mixing of symbols and strings is also possible.

cell(@click("rowclick", [:addTableIndex, "console.log('Click coords: ' + event.clientX + '|' + event.clientY)", :myHandler]))

Stipple.Elements.@showif — Macro

@showif(expr, [type])

v-show will always be rendered and remain in the DOM; v-show only toggles the display CSS property of the element. https://vuejs.org/v2/guide/conditional.html#v-show

Difference between @showif and @iif when to use either

v-if has higher toggle costs while v-show has higher initial render costs

Example

julia> h1("Hello!", @showif(:ok))
"<h1 v-show="ok">Hello!</h1>"

Stipple.Elements.stylesheet — Function

function stylesheet(href::String; args...) :: String

Generates the corresponding HTML link tag to reference the CSS stylesheet at href.

Example

julia> stylesheet("https://fonts.googleapis.com/css?family=Material+Icons")
"<link href="https://fonts.googleapis.com/css?family=Material+Icons" rel="stylesheet" />"